Collection now returns just the operation instead of [res, op]
During tests was needed to do res, op = Thing::Index.collection everytime, and res is never used
because makes no sense to check if getting a collection was a valid operation, its hard to define
what is a valid search or not. So decided to avoid this behavior and return just the collection.

Was really missing the call syntax during the tests :cry: hahahah