Set right variable name when using Controller::AR
When using Controller::ActiveRecord it tries to set the variable name
according to table name, but when we are using collection there is no
way of doing this automatically because @model is a
ActiveRecord_Relation so what is proposed is to check if @model is a
relation then set the right variable name.

This case was not supported on first collection commit because wasn't
being tested the case that a controller include AR module, now its being
covered and with tests.