Controller#run and friends will not change params anymore unless you say so. this used to happen for formats != :html. You can configure that now.
class CommentsController < ..
  operation document_formats: :json

This will provide the original incoming JSON document in params["comment"] and work seamlessly with a JSON operation that uses a representer. this basically means you have to parse it yourself, so only do this when you want to use a representer in your operation (recommended) or your own parsing.